render: fix underline/strikeout positioning
There were two errors:

* We subtracted half the line width instead of adding it to the
  baseline

* We rounded the line positioning and thickness before the positioning
  calculation. In particular, rounding the thickness before using it
  to adjust the position was wrong. Now we round just before the
  pixman call.

+++ font.h
@@ -53,13 +53,13 @@ struct font {
     int max_x_advance;
 
     struct {
-        int position;
-        int thickness;
+        double position;
+        double thickness;
     } underline;
 
     struct {
-        int position;
-        int thickness;
+        double position;
+        double thickness;
     } strikeout;
 
     bool is_fallback;
